% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepop.R
\name{basepop}
\alias{basepop}
\title{Function to generate baseline population matrix.}
\usage{
basepop(
  inputsPath,
  regUAll,
  yearStart = 2010,
  yearEnd = 2100,
  num.ages = 101,
  cur.scenario = "Constant_rate",
  scen.factor = 1,
  int.mig = 0,
  gen.output = NULL
)
}
\arguments{
\item{inputsPath}{Input folder containing state subdirectories}

\item{regUAll}{Array containing state subfolder names}

\item{yearStart}{Start year for data processing}

\item{yearEnd}{Through year for data processing}

\item{num.ages}{Number of ages (come back and fix this)}

\item{cur.scenario}{Current scenario rate (constant rate, etc.)}

\item{scen.factor}{Target population scenario}

\item{int.mig}{International migration rate (boolean, False = not applied)}

\item{gen.output}{Optional, supply file path+name to generate output csv.}
}
\value{
DataFrame containing baseline population matrix
}
\description{
Function to generate baseline population matrix.
}
