% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility.R
\name{fertility}
\alias{fertility}
\title{Function to generate fertility matrix.}
\usage{
fertility(
  inputsPath,
  regUAll,
  yearStart = 2010,
  yearEnd = 2100,
  cur.scenario = "Constant_rate",
  gen.output = NULL,
  useBrassf = T,
  vis = T
)
}
\arguments{
\item{regUAll}{Array containing state subfolder names}

\item{yearStart}{Start year for data processing}

\item{yearEnd}{Through year for data processing}

\item{cur.scenario}{Target population scenario}

\item{gen.output}{Optional, supply file path+name to generate output csv.}

\item{useBrassf}{(Boolean) Should the Brass Relational Model be used or a simple scaling approach to compute future fertility schedules}

\item{vis}{(Boolean) Should details for projection model adjustment be printed?}

\item{pathIn}{Input folder containing state subdirectories}
}
\value{
DataFrame containing fertility matrix
}
\description{
Function to generate fertility matrix.
}
