% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality}
\alias{mortality}
\title{Function to generate mortality matrices.}
\usage{
mortality(pathIn, regUAll, cur.scenario = "Constant_rate", gen.output = NULL)
}
\arguments{
\item{pathIn}{Input folder containing state subdirectories}

\item{regUAll}{Array containing state subfolder names}

\item{cur.scenario}{Target population scenario}

\item{generate.output}{Optional, supply file path+name to generate output csv.}
}
\value{
Mortality matrix (1).
}
\description{
Function to generate mortality matrices.
}
