% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_projection.R
\name{pop.projection}
\alias{pop.projection}
\title{Function to project population migration totals}
\usage{
pop.projection(
  inputsPath,
  regUAll,
  yearStart = 2010,
  yearEnd = 2100,
  scen.factor = 0,
  num.ages = 101,
  int.mig = 0,
  cur.scenario = "Constant_rate",
  gen.output = NULL
)
}
\arguments{
\item{inputsPath}{Input folder containing state subdirectories}

\item{regUAll}{Array containing state subfolder names}

\item{yearStart}{Start year for data processing}

\item{yearEnd}{Through year for data processing}

\item{scen.factor}{Target population scenario}

\item{num.ages}{Come back to fix}

\item{int.mig}{International migration rate (boolean, False = not applied)}

\item{cur.scenario}{Current scenario rate (constant rate, etc.)}

\item{generate.output}{Optional, supply file path+name to generate output csv.}
}
\description{
Function to project population migration totals
}
