% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality}
\alias{mortality}
\title{Function to generate mortality matrices.}
\usage{
mortality(
  inputsPath,
  regUAll,
  datMortS30S,
  datMortS100S,
  yearStart = 2010,
  yearEnd = 2100,
  cur.scenario = "Constant_rate",
  gen.output = NULL,
  vis = F
)
}
\arguments{
\item{inputsPath}{Input folder containing state subdirectories}

\item{regUAll}{Array containing state subfolder names}

\item{datMortS30S}{UN Standard 30-year life table}

\item{datMortS100S}{UN Standard 100-year life table}

\item{yearStart}{Start year for data processing}

\item{yearEnd}{Through year for data processing}

\item{cur.scenario}{Current scenario rate (constant rate, etc.)}

\item{gen.output}{Optional, supply file path+name to generate output csv.}

\item{vis}{(Boolean) Should details for projection model adjustment be printed?}
}
\value{
Mortality matrix (1).
}
\description{
Function to generate mortality matrices.
}
